# OpenApiGeneratorPlayground/Generator/Output/MyPackageClient/Model/JobStatus.cs
// <auto-generated>
/*
 * Job Processing API
 *
 * API for submitting and tracking background jobs.
 *
 * The version of the OpenAPI document: 1.0.0
 * Generated by: https://github.com/openapitools/openapi-generator.git
 */

#nullable enable

using System;
using System.Collections;
using System.Collections.Generic;
using System.Collections.ObjectModel;
using System.Linq;
using System.IO;
using System.Text;
using System.Text.RegularExpressions;
using System.Text.Json;
using System.Text.Json.Serialization;
using System.ComponentModel.DataAnnotations;
using MyPackageClient.Client;

namespace MyPackageClient.Model
{
    /// <summary>
    /// Defines JobStatus
    /// </summary>
    public enum JobStatus
    {
        /// <summary>
        /// Enum Queued for value: queued
        /// </summary>
        Queued = 1,

        /// <summary>
        /// Enum Running for value: running
        /// </summary>
        Running = 2,

        /// <summary>
        /// Enum Completed for value: completed
        /// </summary>
        Completed = 3,

        /// <summary>
        /// Enum Failed for value: failed
        /// </summary>
        Failed = 4
    }

    /// <summary>
    /// Converts <see cref="JobStatus"/> to and from the JSON value
    /// </summary>
    public static class JobStatusValueConverter
    {
        /// <summary>
        /// Parses a given value to <see cref="JobStatus"/>
        /// </summary>
        /// <param name="value"></param>
        /// <returns></returns>
        public static JobStatus FromString(string value)
        {
            if (value.Equals("queued"))
                return JobStatus.Queued;

            if (value.Equals("running"))
                return JobStatus.Running;

            if (value.Equals("completed"))
                return JobStatus.Completed;

            if (value.Equals("failed"))
                return JobStatus.Failed;

            throw new NotImplementedException($"Could not convert value to type JobStatus: '{value}'");
        }

        /// <summary>
        /// Parses a given value to <see cref="JobStatus"/>
        /// </summary>
        /// <param name="value"></param>
        /// <returns></returns>
        public static JobStatus? FromStringOrDefault(string value)
        {
            if (value.Equals("queued"))
                return JobStatus.Queued;

            if (value.Equals("running"))
                return JobStatus.Running;

            if (value.Equals("completed"))
                return JobStatus.Completed;

            if (value.Equals("failed"))
                return JobStatus.Failed;

            return null;
        }

        /// <summary>
        /// Converts the <see cref="JobStatus"/> to the json value
        /// </summary>
        /// <param name="value"></param>
        /// <returns></returns>
        /// <exception cref="NotImplementedException"></exception>
        public static string ToJsonValue(JobStatus value)
        {
            if (value == JobStatus.Queued)
                return "queued";

            if (value == JobStatus.Running)
                return "running";

            if (value == JobStatus.Completed)
                return "completed";

            if (value == JobStatus.Failed)
                return "failed";

            throw new NotImplementedException($"Value could not be handled: '{value}'");
        }
    }

    /// <summary>
    /// A Json converter for type <see cref="JobStatus"/>
    /// </summary>
    /// <exception cref="NotImplementedException"></exception>
    public class JobStatusJsonConverter : JsonConverter<JobStatus>
    {
        /// <summary>
        /// Returns a  from the Json object
        /// </summary>
        /// <param name="reader"></param>
        /// <param name="typeToConvert"></param>
        /// <param name="options"></param>
        /// <returns></returns>
        public override JobStatus Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
        {
            string? rawValue = reader.GetString();

            JobStatus? result = rawValue == null
                ? null
                : JobStatusValueConverter.FromStringOrDefault(rawValue);

            if (result != null)
                return result.Value;

            throw new JsonException();
        }

        /// <summary>
        /// Writes the JobStatus to the json writer
        /// </summary>
        /// <param name="writer"></param>
        /// <param name="jobStatus"></param>
        /// <param name="options"></param>
        public override void Write(Utf8JsonWriter writer, JobStatus jobStatus, JsonSerializerOptions options)
        {
            writer.WriteStringValue(JobStatusValueConverter.ToJsonValue(jobStatus).ToString());
        }
    }

    /// <summary>
    /// A Json converter for type <see cref="JobStatus"/>
    /// </summary>
    public class JobStatusNullableJsonConverter : JsonConverter<JobStatus?>
    {
        /// <summary>
        /// Returns a JobStatus from the Json object
        /// </summary>
        /// <param name="reader"></param>
        /// <param name="typeToConvert"></param>
        /// <param name="options"></param>
        /// <returns></returns>
        public override JobStatus? Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
        {
            string? rawValue = reader.GetString();

            JobStatus? result = rawValue == null
                ? null
                : JobStatusValueConverter.FromStringOrDefault(rawValue);

            if (result != null)
                return result.Value;

            throw new JsonException();
        }

        /// <summary>
        /// Writes the JobStatus to the json writer
        /// </summary>
        /// <param name="writer"></param>
        /// <param name="jobStatus"></param>
        /// <param name="options"></param>
        public override void Write(Utf8JsonWriter writer, JobStatus? jobStatus, JsonSerializerOptions options)
        {
            writer.WriteStringValue(jobStatus.HasValue ? JobStatusValueConverter.ToJsonValue(jobStatus.Value).ToString() : "null");
        }
    }
}