# OpenApiGeneratorPlayground/Generator/Output/MyPackageServer/Controllers/JobsApi.cs
/*
 * Job Processing API
 *
 * API for submitting and tracking background jobs.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.Http;
using MyPackageServer.Models;

namespace MyPackageServer.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public abstract class JobsApiController : ControllerBase
    { 
        /// <summary>
        /// Get job by ID
        /// </summary>
        /// <param name="jobId"></param>
        /// <response code="200">Job details</response>
        /// <response code="404">Resource not found</response>
        [HttpGet]
        [Route("/jobs/{jobId}")]
        [ProducesResponseType(statusCode: 200, type: typeof(Job))]
        [ProducesResponseType(statusCode: 404, type: typeof(Error))]
        public abstract IActionResult GetJob([FromRoute (Name = "jobId")][Required]Guid jobId);

        /// <summary>
        /// List jobs
        /// </summary>
        /// <param name="limit"></param>
        /// <param name="offset"></param>
        /// <response code="200">List of jobs</response>
        [HttpGet]
        [Route("/jobs")]
        [ProducesResponseType(statusCode: 200, type: typeof(List<Job>))]
        public abstract IActionResult ListJobs([FromQuery (Name = "limit")][Range(1, 100)]int? limit, [FromQuery (Name = "offset")]int? offset);

        /// <summary>
        /// Submit a job
        /// </summary>
        /// <param name="jobRequest"></param>
        /// <response code="201">Job created</response>
        /// <response code="400">Invalid input</response>
        [HttpPost]
        [Route("/jobs")]
        [Consumes("application/json")]
        [ProducesResponseType(statusCode: 201, type: typeof(Job))]
        [ProducesResponseType(statusCode: 400, type: typeof(Error))]
        public abstract IActionResult SubmitJob([FromBody]JobRequest jobRequest);
    }
}