# OpenApiGeneratorPlayground/Generator/Output/MyPackageServer/Models/DataExportJob.cs
/*
 * Job Processing API
 *
 * API for submitting and tracking background jobs.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Linq;
using System.Text;
using System.Collections.Generic;
using System.ComponentModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using System.Text.Json;

namespace MyPackageServer.Models
{ 
    /// <summary>
    /// 
    /// </summary>
    [DataContract]
    public class DataExportJob 
    {

        /// <summary>
        /// Gets or Sets Format
        /// </summary>
        
        public enum FormatEnum
        {
            
            /// <summary>
            /// Enum CsvEnum for csv
            /// </summary>
            [EnumMember(Value = "csv")]
            CsvEnum = 1,
            
            /// <summary>
            /// Enum JsonEnum for json
            /// </summary>
            [EnumMember(Value = "json")]
            JsonEnum = 2
        }

        /// <summary>
        /// Gets or Sets Format
        /// </summary>
        [DataMember(Name="format", EmitDefaultValue=true)]
        public FormatEnum Format { get; set; }

        /// <summary>
        /// Gets or Sets DownloadUrl
        /// </summary>
        [DataMember(Name="downloadUrl", EmitDefaultValue=false)]
        public string? DownloadUrl { get; set; } = null!;

    }
}