# OpenApiGeneratorPlayground/Generator/Output/MyPackageServer/Models/DataExportJobRequest.cs
/*
 * Job Processing API
 *
 * API for submitting and tracking background jobs.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Linq;
using System.Text;
using System.Collections.Generic;
using System.ComponentModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using System.Text.Json;

namespace MyPackageServer.Models
{ 
    /// <summary>
    /// 
    /// </summary>
    [DataContract]
    public class DataExportJobRequest 
    {

        /// <summary>
        /// Gets or Sets Type
        /// </summary>
        
        public enum TypeEnum
        {
            
            /// <summary>
            /// Enum DataExportEnum for dataExport
            /// </summary>
            [EnumMember(Value = "dataExport")]
            DataExportEnum = 1
        }

        /// <summary>
        /// Gets or Sets Type
        /// </summary>
        [Required]
        [DataMember(Name="type", EmitDefaultValue=true)]
        public required TypeEnum Type { get; set; }


        /// <summary>
        /// Gets or Sets Format
        /// </summary>
        
        public enum FormatEnum
        {
            
            /// <summary>
            /// Enum CsvEnum for csv
            /// </summary>
            [EnumMember(Value = "csv")]
            CsvEnum = 1,
            
            /// <summary>
            /// Enum JsonEnum for json
            /// </summary>
            [EnumMember(Value = "json")]
            JsonEnum = 2
        }

        /// <summary>
        /// Gets or Sets Format
        /// </summary>
        [Required]
        [DataMember(Name="format", EmitDefaultValue=true)]
        public required FormatEnum Format { get; set; }

        /// <summary>
        /// Gets or Sets IncludeMetadata
        /// </summary>
        [DataMember(Name="includeMetadata", EmitDefaultValue=true)]
        public bool? IncludeMetadata { get; set; } = false;

    }
}