# OpenApiGeneratorPlayground/Generator/Output/MyPackageServer/Models/EmailJob.cs
/*
 * Job Processing API
 *
 * API for submitting and tracking background jobs.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Linq;
using System.Text;
using System.Collections.Generic;
using System.ComponentModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using System.Text.Json;

namespace MyPackageServer.Models
{ 
    /// <summary>
    /// 
    /// </summary>
    [DataContract]
    public class EmailJob 
    {
        /// <summary>
        /// Gets or Sets Recipient
        /// </summary>
        [DataMember(Name="recipient", EmitDefaultValue=false)]
        public string? Recipient { get; set; } = null!;

        /// <summary>
        /// Gets or Sets Subject
        /// </summary>
        [DataMember(Name="subject", EmitDefaultValue=false)]
        public string? Subject { get; set; } = null!;

    }
}