# OpenApiGeneratorPlayground/Generator/Output/MyPackageServer/Models/EmailJobRequest.cs
/*
 * Job Processing API
 *
 * API for submitting and tracking background jobs.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Linq;
using System.Text;
using System.Collections.Generic;
using System.ComponentModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using System.Text.Json;

namespace MyPackageServer.Models
{ 
    /// <summary>
    /// 
    /// </summary>
    [DataContract]
    public class EmailJobRequest 
    {

        /// <summary>
        /// Gets or Sets Type
        /// </summary>
        
        public enum TypeEnum
        {
            
            /// <summary>
            /// Enum EmailEnum for email
            /// </summary>
            [EnumMember(Value = "email")]
            EmailEnum = 1
        }

        /// <summary>
        /// Gets or Sets Type
        /// </summary>
        [Required]
        [DataMember(Name="type", EmitDefaultValue=true)]
        public required TypeEnum Type { get; set; }

        /// <summary>
        /// Gets or Sets Recipient
        /// </summary>
        [Required]
        [DataMember(Name="recipient", EmitDefaultValue=false)]
        public required string Recipient { get; set; }

        /// <summary>
        /// Gets or Sets Subject
        /// </summary>
        [Required]
        [DataMember(Name="subject", EmitDefaultValue=false)]
        public required string Subject { get; set; }

        /// <summary>
        /// Gets or Sets Body
        /// </summary>
        [DataMember(Name="body", EmitDefaultValue=false)]
        public string? Body { get; set; } = null!;

    }
}