# OpenApiGeneratorPlayground/Generator/Output/MyPackageServer/Models/EmailTarget.cs
/*
 * Job Processing API
 *
 * API for submitting and tracking background jobs.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Linq;
using System.Text;
using System.Collections.Generic;
using System.ComponentModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using System.Text.Json;

namespace MyPackageServer.Models
{ 
    /// <summary>
    /// 
    /// </summary>
    [DataContract]
    public class EmailTarget 
    {
        /// <summary>
        /// Gets or Sets Email
        /// </summary>
        [Required]
        [DataMember(Name="email", EmitDefaultValue=false)]
        public required string Email { get; set; }

    }
}