# OpenApiGeneratorPlayground/Generator/Output/MyPackageServer/Models/Job.cs
/*
 * Job Processing API
 *
 * API for submitting and tracking background jobs.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Linq;
using System.Text;
using System.Collections.Generic;
using System.ComponentModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using System.Text.Json;
using JsonSubTypes;

namespace MyPackageServer.Models
{ 
    /// <summary>
    /// 
    /// </summary>
    [DataContract]
    [JsonSubtypes.KnownSubType(typeof(DataExportJob), "dataExport")]
    [JsonSubtypes.KnownSubType(typeof(EmailJob), "email")]
    public class Job 
    {
        /// <summary>
        /// Gets or Sets Id
        /// </summary>
        [Required]
        [DataMember(Name="id", EmitDefaultValue=true)]
        public required Guid Id { get; set; }

        /// <summary>
        /// Gets or Sets Type
        /// </summary>
        [Required]
        [DataMember(Name="type", EmitDefaultValue=false)]
        public required string Type { get; set; }

        /// <summary>
        /// Gets or Sets Status
        /// </summary>
        [Required]
        [DataMember(Name="status", EmitDefaultValue=true)]
        public required JobStatus Status { get; set; }

        /// <summary>
        /// Gets or Sets CreatedAt
        /// </summary>
        [Required]
        [DataMember(Name="createdAt", EmitDefaultValue=true)]
        public required DateTime CreatedAt { get; set; }

        /// <summary>
        /// Gets or Sets CompletedAt
        /// </summary>
        [DataMember(Name="completedAt", EmitDefaultValue=true)]
        public DateTime? CompletedAt { get; set; } = null!;

        /// <summary>
        /// Gets or Sets Recipient
        /// </summary>
        [DataMember(Name="recipient", EmitDefaultValue=false)]
        public string? Recipient { get; set; } = null!;

        /// <summary>
        /// Gets or Sets Subject
        /// </summary>
        [DataMember(Name="subject", EmitDefaultValue=false)]
        public string? Subject { get; set; } = null!;


        /// <summary>
        /// Gets or Sets Format
        /// </summary>
        
        public enum FormatEnum
        {
            
            /// <summary>
            /// Enum CsvEnum for csv
            /// </summary>
            [EnumMember(Value = "csv")]
            CsvEnum = 1,
            
            /// <summary>
            /// Enum JsonEnum for json
            /// </summary>
            [EnumMember(Value = "json")]
            JsonEnum = 2
        }

        /// <summary>
        /// Gets or Sets Format
        /// </summary>
        [DataMember(Name="format", EmitDefaultValue=true)]
        public FormatEnum Format { get; set; }

        /// <summary>
        /// Gets or Sets DownloadUrl
        /// </summary>
        [DataMember(Name="downloadUrl", EmitDefaultValue=false)]
        public string? DownloadUrl { get; set; } = null!;

    }
}